% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbahadur_unstructured.R
\name{rbahadur_unstructured}
\alias{rbahadur_unstructured}
\title{Binary random variates with unstructured correlations}
\usage{
rbahadur_unstructured(n, mu, C)
}
\arguments{
\item{n}{number of observations}

\item{mu}{vector of means}

\item{C}{correlation matrix}
}
\value{
An \eqn{n}-by-\eqn{m} matrix of binary random variates, where \eqn{m} is the length of 'mu'.
}
\description{
Generate second Bahadur order multivariate Bernoulli random variates with unstructured correlations.
}
\details{
This generates multivariate Bernoulli (MVB) random vectors with mean vector 'mu' and correlation matrix 'C'. 'mu' must take values in the open unit interval and 'C' must induce a valid second Bahadur order probability distribution. That is, there must exist an MVB probability distribution with first moments 'mu' and standardized central second moments 'C' such that all higher order central moments are zero.
}
