% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_covariance_structure.R
\name{am_covariance_structure}
\alias{am_covariance_structure}
\title{Compute Diagonal plus Low Rank covariance structure under equilibrium assortative mating}
\usage{
am_covariance_structure(beta, AF, r)
}
\arguments{
\item{beta}{vector of standardized diploid allele-substitution effects}

\item{AF}{vector of allele frequencies}

\item{r}{cross-mate phenotypic correlation}
}
\value{
Vector 'U' such that $D + U U^T$ corresponds to the expected haploid LD-matrix given the specified genetic architecture (encoded by 'beta' and 'AF') and cross-mate phenotypic correlation 'r'. It is assumed that the total phenotypic variance at generation zero is one.
}
\description{
Compute Diagonal plus Low Rank covariance structure under equilibrium assortative mating
}
