% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_simulate.R
\name{am_simulate}
\alias{am_simulate}
\title{Simulate genotype/phenotype data under univariate AM}
\usage{
am_simulate(h2_0, r, m, n, min_MAF = 0.1)
}
\arguments{
\item{h2_0}{generation zero (panmictic) heritability}

\item{r}{cross-mate phenotypic correlation}

\item{m}{number of biallelic causal variants}

\item{n}{sample size}

\item{min_MAF}{minimum minor allele frequency for causal variants}
}
\value{

}
\description{
Simulate genotype/phenotype data under univariate AM
}
\examples{
set.seed(1)
h2_0 = .5; m = 200; n = 5000; r =.5

## simulate genotype/phenotype data
sim_dat <- am_simulate(h2_0, r, m, n)
str(sim_dat)

## empirical h2 vs expected equilibrium h2
(emp_h2 <- var(sim_dat$g)/var(sim_dat$y))
h2_eq(r, h2_0)
}
