% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_simulate.R
\name{am_simulate}
\alias{am_simulate}
\title{Simulate genotype/phenotype data under univariate AM}
\usage{
am_simulate(h2_0, r, m, n, min_MAF = 0.1)
}
\arguments{
\item{h2_0}{generation zero (panmictic) heritability}

\item{r}{cross-mate phenotypic correlation}

\item{m}{number of biallelic causal variants}

\item{n}{sample size}

\item{min_MAF}{minimum minor allele frequency for causal variants}
}
\value{

}
\description{
Simulate genotype/phenotype data under univariate AM
}
